pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "altair==6.0.0",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==25.1.0",
    "arrow==1.3.0",
    "asgiref==3.11.1",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "attrs==25.3.0",
    "audioop-lts==0.2.2",
    "azure-ai-contentsafety==1.0.0",
    "azure-ai-textanalytics==5.3.0",
    "azure-ai-vision-face==1.0.0b2",
    "azure-ai-vision-imageanalysis==1.0.0",
    "azure-common==1.1.28",
    "azure-core==1.38.1",
    "azure-core-tracing-opentelemetry==1.0.0b12",
    "azure-identity==1.25.2",
    "azure-monitor-opentelemetry==1.8.6",
    "azure-monitor-opentelemetry-exporter==1.0.0b48",
    "azure-search-documents==11.6.0",
    "azure-storage-blob==12.28.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.5",
    "bleach==6.2.0",
    "blinker==1.9.0",
    "cachetools==6.2.6",
    "certifi==2025.8.3",
    "cffi==2.0.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "comm==0.2.2",
    "comtypes==1.4.11",
    "contourpy==1.3.3",
    "cryptography==46.0.5",
    "curl-cffi==0.13.0",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "distlib==0.4.0",
    "distro==1.9.0",
    "executing==2.2.0",
    "fastapi==0.135.1",
    "fastjsonschema==2.21.2",
    "filelock==3.19.1",
    "firecrawl-py==4.18.1",
    "flask==3.1.1",
    "fonttools==4.60.0",
    "fqdn==1.5.1",
    "frozendict==2.4.6",
    "frozenlist==1.8.0",
    "gitdb==4.0.12",
    "gitpython==3.1.46",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.10",
    "importlib-metadata==8.7.1",
    "ipykernel==6.29.5",
    "ipython==9.3.0",
    "ipython-pygments-lexers==1.1.1",
    "isodate==0.7.2",
    "isoduration==20.11.0",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "joblib==1.5.2",
    "json5==0.12.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.9.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.3.0",
    "jupyter-server==2.17.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.7",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "kiwisolver==1.4.9",
    "langchain==1.2.10",
    "langchain-classic==1.0.2",
    "langchain-community==0.4.1",
    "langchain-core==1.2.18",
    "langchain-openai==1.1.11",
    "langchain-text-splitters==1.1.1",
    "langgraph==1.0.10",
    "langgraph-checkpoint==4.0.1",
    "langgraph-prebuilt==1.0.8",
    "langgraph-sdk==0.3.10",
    "langsmith==0.7.16",
    "lark==1.3.0",
    "lxml==6.0.2",
    "markupsafe==3.0.2",
    "marshmallow==3.26.2",
    "matplotlib==3.10.6",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.4",
    "msal==1.35.1",
    "msal-extensions==1.3.1",
    "msrest==0.7.1",
    "multidict==6.7.1",
    "multitasking==0.0.12",
    "mypy-extensions==1.1.0",
    "narwhals==2.16.0",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "notebook==7.4.5",
    "notebook-shim==0.2.4",
    "numpy==2.3.3",
    "oauthlib==3.3.1",
    "openai==2.26.0",
    "opentelemetry-api==1.39.0",
    "opentelemetry-instrumentation==0.60b0",
    "opentelemetry-instrumentation-asgi==0.60b0",
    "opentelemetry-instrumentation-dbapi==0.60b0",
    "opentelemetry-instrumentation-django==0.60b0",
    "opentelemetry-instrumentation-fastapi==0.60b0",
    "opentelemetry-instrumentation-flask==0.60b0",
    "opentelemetry-instrumentation-psycopg2==0.60b0",
    "opentelemetry-instrumentation-requests==0.60b0",
    "opentelemetry-instrumentation-urllib==0.60b0",
    "opentelemetry-instrumentation-urllib3==0.60b0",
    "opentelemetry-instrumentation-wsgi==0.60b0",
    "opentelemetry-resource-detector-azure==0.1.5",
    "opentelemetry-sdk==1.39.0",
    "opentelemetry-semantic-conventions==0.60b0",
    "opentelemetry-util-http==0.60b0",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==25.0",
    "pandas==2.3.2",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "peewee==3.18.2",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "prometheus-client==0.23.1",
    "prompt-toolkit==3.0.51",
    "propcache==0.4.1",
    "protobuf==6.32.1",
    "psutil==7.0.0",
    "psycopg2-binary==2.9.11",
    "pure-eval==0.2.3",
    "pyarrow==23.0.0",
    "pyaudio==0.2.14",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.13.1",
    "pydeck==0.9.1",
    "pygments==2.19.2",
    "pyjwt==2.11.0",
    "pyparsing==3.2.5",
    "pypdf==6.8.0",
    "pypiwin32==223",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "python-json-logger==3.3.0",
    "pyttsx3==2.99",
    "pytz==2025.2",
    "pywin32==310",
    "pywinpty==3.0.0",
    "pyyaml==6.0.2",
    "pyzmq==27.0.0",
    "redis==7.3.0",
    "referencing==0.36.2",
    "regex==2026.2.28",
    "requests==2.32.5",
    "requests-oauthlib==2.0.0",
    "requests-toolbelt==1.0.0",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rfc3987-syntax==1.1.0",
    "rpds-py==0.27.1",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "send2trash==1.8.3",
    "setuptools==80.9.0",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.8",
    "speechrecognition==3.14.3",
    "sqlalchemy==2.0.48",
    "stack-data==0.6.3",
    "standard-aifc==3.13.0",
    "standard-chunk==3.13.0",
    "starlette==0.52.1",
    "streamlit==1.54.0",
    "tenacity==9.1.4",
    "terminado==0.18.1",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tinycss2==1.4.0",
    "toml==0.10.2",
    "tornado==6.5.1",
    "tqdm==4.67.3",
    "traitlets==5.14.3",
    "types-python-dateutil==2.9.0.20250822",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "uri-template==1.3.0",
    "urllib3==2.5.0",
    "uuid-utils==0.14.1",
    "uv==0.10.9",
    "uvicorn==0.41.0",
    "virtualenv==20.34.0",
    "watchdog==6.0.0",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "wrapt==1.17.3",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "yfinance==0.2.66",
    "yt-dlp==2026.3.3",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
